+++ app/src/main/java/nextstep/shoppingcart/Presentation/MainActivity.kt
@@ -1,0 +1,17 @@
+package nextstep.shoppingcart.Presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import nextstep.shoppingcart.Presentation.ui.shopping.ProductsScreen
+import nextstep.shoppingcart.Presentation.ui.theme.ShoppingCartTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            ShoppingCartTheme {
+            }
+        }
+    }
+}
