+++ app/src/main/java/nextstep/shoppingcart/Presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.ui.theme
+package nextstep.shoppingcart.Presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
