+++ app/src/main/java/nextstep/shoppingcart/Presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.ui.theme
+package nextstep.shoppingcart.Presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
